options/posix: move nl_langinfo to internal for internal use
This is necessary e.g. for strftime, as it prints strings that are
locale-aware, but nl_langinfo is part of the posix option. In order to
resolve this, we're moving the implementation of nl_langinfo into
internal, making the posix option implementation a wrapper around that.
With this, we also need to move the definition of nl_item and its values
out. This is achieved by putting it all into the nl_item bits header,
that is moved into internal as well.