sysdeps/linux: generate complete syscall lists instead
Our manually curated syscall lists are prone to errors, are a bother to
manually write, and might not operate well on real systems (due to being
incomplete, typo'd, or similar). Instead of this manual and error-prone
process, we can automate fetching complete and accurate syscall lists
straight from the kernel itself.

The script that does this job is sysdeps/linux/update-syscall-list.py.
It fetches kernel sources, unpacks them, discovers all the architectures
it provides, and runs a set of harnesses on them (a harness consists of
a build configuration and an architecture). It then invokes GCC to
provide a list of all the defines in the asm/unistd.h header (and it's
transitive includes), and collects them into a global set, as well as a
local, per-harness set, that's used to build up a program that gives us
all relevant syscall numbers, for use during building. The latter part
of this process is done to avoid requiring linux-headers to build mlibc,
as that would require developers and users of mlibc on Linux to go
through the slight inconvenience of getting a cross toolchain for mlibc.

The list of relevant syscall numbers is generated by generating a C
program that printfs syscalls and their numbers as a JSON dictionary,
letting the compiler handle the job of parsing the syscall definition
macros (which are often not just a number, and invoke some bitwise or
arithmetic operators, such as for x32). Now that we have a dictionary
with all the relevant syscalls, we write this list out, in number order,
as sysdeps/linux/$ARCH/syscallnos.h, which is only used internally
through sysdeps/linux/include-internal/linux/unistd.h, a stub header
that is also not installed, and redirects to syscallnos.h, which
serves to emulate real systems (and is separate enough to be
possibly omitted in the future, if some hypothetical flag such as
-Duse_standard_includes=true is passed to the build system).

The global set of syscalls is used to generate a header that is
installed: sysdeps/linux/include/bits/syscall_aliases.h. It contains a
list of syscall names and stubs to generate SYS_<name> aliases for them,
which are used by some programs like sys-apps/attr. This header is
invoked by <sys/syscall.h> after including <linux/unistd.h>.