unistd.h: provide bits/syscall.h only if platform has it
Currently, the only platform providing bits/syscall.h is Linux. Linux is
also the only platform whose users rely on unistd.h including
<bits/syscall.h>. Precedent set by other platforms is that the syscall
function is provided by the sysdep itself, and as such, this would not
be a useful change.

In conclusion, <bits/syscall.h> is provided on an as-wants basis,
depending on platform.