options/rtdl: set object->localScope to globalScope for global objects
This fixes a crash that if an object was already in the global scope
(as a result of for example DT_NEEDED), dlopening it again with
RTLD_LOCAL would try to use object->localScope, which did not point to
anything, while the object should remain in the global scope instead.

Also added a regression test for this.