scripts: add a script to generate rust bindings for linux_like targets
This script and the configuration file supplied are tested to generate
valid rust bindings (for rust's libc crate) for unix/linux_like/*
targets. This has been tested by building rust's stdlib, alacritty, exa
and ripgrep.

In the future, it might be a worthwhile idea to run this as a GHA check
to catch accidental breakage.