options/ansi: fix UAF for tzname
`tzname` points into static storage, but previously it was set to point
into a file_window, which gets dropped when it goes out of scope. As an
alternate strategy, we use two frigg strings as static storage, so that
we can copy out the string values from the window before unmapping.